requirements.txt
fastapi
uvicorn[standard]
jinja2
openai
python-dotenv
fastapi
jinja2
openai
python-dotenv
uvicorn[standard]
python-multipart
fpdf2
beautifulsoup4
httpx
python-multipart
sqlalchemy
itsdangerous
starlette
stripe
bcrypt
passlib
google-generativeai
